work on analyzer
some fairly significant changes that will result in more jobs erroring out as a result of testing on a mac pro.  first, if a job doesn't produce a frame for analysis at least once every 10 seconds, it's determined to have hung, and will be cancelled with an error.  second, the number of sample buffers retrieved from the analysis output is compared to the number of sample buffers retrieved from the passthru output- if there's a difference, the file is flagged as an error... :: also did a little work to make sure the app functions correctly if using a preset that has disabled analysis